Textbased Dungeon Game with C
The game consists of 12 rooms. Starting from the first room, we encounter various monsters, wander around the rooms and collect items. Each item has its own properties and these have an effect. We can navigate between rooms with commands. With the commands, we can kill monsters, collect items, save and load. With the Look command, we see the room. With the Inventory command, we see the items in our inventory.